(Makefile.PL): Correct required version of Test::More
According to the Perl documentation, the functions note() and
explain() appeared in Test::More version 0.82. Since these functions
are used within the DBD::SQLite tests, the minimum required version
of Test::More needs to be corrected. Without the proper version,
the build fails on Perls earlier than 5.10. For more information:
https://perldoc.perl.org/Test/More.html#COMPATIBILITY