Die on EOF() method mismatch
perl_vt_Eof() in dbdimp_virtual_table.inc calls EOF() method. If the
call returned an unexpected number of values, perl_vt_Eof() printed
a warning and returned a value of a local variable that was never
initialized.

This patch fixes it by raising an exception instead of printing
a warning.