Give upgraded strings to SQLite as downgraded when sqlite_unicode==0.
Issue #66: If $a and $b contain the same code points but one is upgraded
and the other isn’t, previously DBD::SQLite would give two different C
strings to SQLite. This changeset fixes that so that SQLite always
translates equivalent Perl strings to the same C string, regardless of the
Perl strings’ internal encoding.